+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes
+- Fix all kind of issues related to font-faces ([#1838](https://github.com/maplibre/maplibre-style-spec/pull/1838))
 - _...Add new stuff here..._
 
 ## 26.4.0